device_management/requirements.txt
flask==2.2.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
paho-mqtt==1.5.1

flask-session==0.4.0
flasgger==0.9.7.1  # Instead of 0.9.5

redis==5.2.0
werkzeug==3.0.2
pika==1.2.0

